+++ modules/tinymce/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Removed
+- Removed deprecated `Schema` settings #TINY-7821
+
 ## 5.10.0 - TBD
 
 ### Added